requirements.txt
attrs==21.2.0
autopep8==1.5.7
click==8.0.1
cursor==1.3.4
Flask==2.0.1
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
toml==0.10.2
Werkzeug==2.0.1
